✨ Allow static Parameters for Quantum Operations (#7)
This PR allows the possibility to define statically known parameters for
quantum operations. Those are given as attributes instead of operands.
For the case of operations that take more than one parameter, there must
be a way to specify which of the parameters is given statically and
which is given dynamically because they cannot be written as a mixed
list in the signature (at least not without writing a custom
parser/printer). For this purpose, we defined a `params_mask` that has
the exact same length as there are parameters. If the value at index i
is true, this means that this parameter is given as an attribute, if it
is false, then this parameter is given as an operand. The attributes and
operands are given in order. Below there is an example of an operation
taking three parameters whereof the first and last are given as operands
and the second as an attribute.

```mlir
%c = arith.constant 0.5 : f64;
%1 = mqtopt.u(%c, %c static [0.3] mask [false, true, false]) %0 : !mqtopt.Qubit
```

---------

Co-authored-by: burgholzer <burgholzer@me.com>