📝 Add documentation for ZX-calculus package and C++ API docs setup (#817)
## Description

This PR introduces documentation for the C++ ZX-calculus library. 
It also adds the general infrastructure for C++ API documentation as
part of the MQT Core project based on Doxygen and the `breathe`
extension.

Fixes #813

This change addresses concerns about missing documentation that came up
in the revision openjournals/joss-reviews#7478.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: Tom Peham <pehamtom@gmx.at>
Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>