📝 Update documentation to reflect recent changes (#861)
## Description

This pull request removes an outdated paragraph from the documentation
and updates the copyright year to the current year. These changes ensure
that the documentation is accurate and up to date.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new