✨Add continuous testing workflow for Qiskit main branch (#860)
## Description

This pull request introduces a continuous testing workflow to ensure
compatibility with the latest changes in the Qiskit main branch. This
addition aims to maintain project stability and address potential
integration issues promptly.
This is especially critical in light of the upcoming Qiskit 2.0 release.

The respective workflow will run once a week, on pull requests that
change the respective file, as well as on manual workflow triggers.

## Checklist:

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>