🎨 Add U operation and allow for Multiple Zones (#877)
## Description

This small PR add a local U3 operation and allows that global operations
can be performed in multiple zones. Additionally, the zone defines an
extent. With that it can be checked whether an atom is contained within
a zone.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>