🔥 Remove `extract_probability_vector` functionality (#883)
## Description

This PR removes the `extract_probability_vector` functionality from
mqt-core in preparation for the v3 launch.
It was only ever used for an evaluation in a paper and never made it
into production code.
Might as well remove it here. Removed code is debugged code 🤷🏼 

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>