🎨 slightly touch up template instantiations and use more trailing return types (#902)
## Description

Just a small byproduct of a debugging session that eventually boiled
down to LLVM@20 and AppleClang@17 not being binary compatible.
Thee PR merely contains syntactic sugar and code style improvements.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: burgholzer <burgholzer@me.com>